# scrivia-v2-openapi-schema.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Scrivia API v2 - LLM Tools",
    "description": "API Scrivia v2 pour l'intégration LLM avec support complet des notes, dossiers et classeurs. Tous les endpoints supportent les slugs et UUIDs.",
    "version": "2.0.0"
  },
  "servers": [
    {
      "url": "https://scrivia.app",
      "description": "Production server"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": { "type": "string" },
          "details": { "type": "array", "items": { "type": "string" } },
          "status": { "type": "integer" }
        },
        "required": ["error"]
      }
    }
  },
  "paths": {
...
    "/api/v2/note/{ref}/table-of-contents": {
      "get": {
        "operationId": "getNoteTableOfContents",
        "summary": "Récupérer la table des matières",
        "description": "Récupérer la table des matières générée automatiquement d'une note",
        "tags": ["Notes"],
        "parameters": [
          {
            "name": "ref",
            "in": "path",
            "required": true,
            "schema": { "type": "string" },
            "description": "ID ou slug de la note"
          }
        ],
        "responses": {
          "200": {
            "description": "Table des matières récupérée avec succès",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "toc": { "type": "array", "items": { "type": "object" } }
                  }
                }
              }
            }
          },
          "404": { "$ref": "#/components/schemas/ErrorResponse" }
        }
      }
    },
...
  }
} 